Changed the regex requirements for Resizetizer. The previous regex "^[a-z]+[a-z0-9_]{0,}[^_]$" doesnt allow single character file names, and also allows a filename such as "a_a="
The new regex follows the error being displayed
First Character must be a letter from a to z
the rest of the string can be a letter from a to z or a number from 0 to 9 and also allows underscores.
Lastly, the last character cannot be an underscore

Those are the three requirements mentioned in the error.
#19574